% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_currency_standards}
\alias{add_currency_standards}
\title{Add currency standards}
\usage{
add_currency_standards(
  Ready4useDyad_r4 = ready4use::Ready4useDyad(),
  default_pkg_ds_chr = c("ISO_4217", "ISOcodes")
)
}
\arguments{
\item{Ready4useDyad_r4}{A dataset and data dictionary pair. (a ready4 S4), Default: ready4use::Ready4useDyad()}

\item{default_pkg_ds_chr}{Default package dataset (a character vector), Default: c("ISO_4217", "ISOcodes")}
}
\value{
A dataset and data dictionary pair. (a ready4 S4)
}
\description{
add_currency_standards() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add currency standards. Function argument Ready4useDyad_r4 specifies the object to be updated. The function returns A dataset and data dictionary pair. (a ready4 S4).
}
\keyword{internal}
