% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_default_country_seed}
\alias{add_default_country_seed}
\title{Add default country seed}
\usage{
add_default_country_seed(CostlySeed_r4 = CostlySeed())
}
\arguments{
\item{CostlySeed_r4}{Original (non-standardised) dataset (and metadata). (a ready4 S4), Default: CostlySeed()}
}
\value{
Original (non-standardised) dataset (and metadata). (a ready4 S4)
}
\description{
add_default_country_seed() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add default country seed. Function argument CostlySeed_r4 specifies the object to be updated. The function returns Original (non-standardised) dataset (and metadata). (a ready4 S4).
}
\keyword{internal}
