% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_default_currency_seed}
\alias{add_default_currency_seed}
\title{Add default currency seed}
\usage{
add_default_currency_seed(
  CostlySeed_r4 = CostlySeed(),
  include_1L_chr = "Currency",
  Ready4useDyad_r4 = ready4use::Ready4useDyad()
)
}
\arguments{
\item{CostlySeed_r4}{Original (non-standardised) dataset (and metadata). (a ready4 S4), Default: CostlySeed()}

\item{include_1L_chr}{Include (a character vector of length one), Default: 'Currency'}

\item{Ready4useDyad_r4}{A dataset and data dictionary pair. (a ready4 S4), Default: ready4use::Ready4useDyad()}
}
\value{
Original (non-standardised) dataset (and metadata). (a ready4 S4)
}
\description{
add_default_currency_seed() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add default currency seed. Function argument CostlySeed_r4 specifies the object to be updated. The function returns Original (non-standardised) dataset (and metadata). (a ready4 S4).
}
\keyword{internal}
