% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_source_label}
\alias{add_source_label}
\title{Add source label}
\usage{
add_source_label(
  CostlySource_r4 = CostlySource(),
  label_1L_chr = "Standardised"
)
}
\arguments{
\item{CostlySource_r4}{Input dataset (and metadata) for generating standardised costing datasets. (a ready4 S4), Default: CostlySource()}

\item{label_1L_chr}{Label (a character vector of length one), Default: 'Standardised'}
}
\value{
Input dataset (and metadata) for generating standardised costing datasets. (a ready4 S4)
}
\description{
add_source_label() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add source label. Function argument CostlySource_r4 specifies the object to be updated. The function returns Input dataset (and metadata) for generating standardised costing datasets. (a ready4 S4).
}
\keyword{internal}
