% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costly-package.R, R/pkg_costly.R
\docType{package}
\name{costly-package}
\alias{costly-package}
\title{costly: Develop, Use and Share Unit Cost Datasets for Health Economic Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking unit costing in conjunction with computational health economic models developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This development version of the costly package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton2@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking unit costing in conjunction with computational health economic models developed with the ready4 framework (https://ready4-dev.github.io/ready4/). This development version of the costly package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton2@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/costly/}
  \item \url{https://github.com/ready4-dev/costly}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/costly/}
  \item \url{https://github.com/ready4-dev/costly}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [funder]

Other contributors:
\itemize{
  \item Monash University [copyright holder, funder]
}

}
\keyword{internal}
