% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_CostlyCountries.R
\docType{class}
\name{CostlyCountries-class}
\alias{CostlyCountries-class}
\alias{CostlyCountries}
\title{CostlyCountries}
\description{
Collection of input, standards definition and results datasets for projects to generate standardised country data for use in costing datasets.
}
\section{Slots}{

\describe{
\item{\code{CostlySeed_r4}}{Original (non-standardised) dataset (and metadata). (a ready4 S4)}

\item{\code{CostlyStandards_r4}}{Dataset (and metadata) defining the allowable values of specified variables. (a ready4 S4)}

\item{\code{fuzzy_logic_1L_chr}}{Fuzzy logic (a character vector of length one)}

\item{\code{correspondences_r3}}{Correspondences (a ready4 S3)}

\item{\code{args_ls}}{Arguments (a list)}

\item{\code{force_standard_1L_lgl}}{Force standard (a logical vector of length one)}

\item{\code{max_distance_1L_dbl}}{Maximum distance (a double vector of length one)}

\item{\code{results_ls}}{Results (a list)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

