% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_CostlySeed.R
\docType{class}
\name{CostlySeed-class}
\alias{CostlySeed-class}
\alias{CostlySeed}
\title{CostlySeed}
\description{
Original (non-standardised) dataset (and metadata).
}
\section{Slots}{

\describe{
\item{\code{args_ls}}{Arguments (a list)}

\item{\code{Ready4useDyad_r4}}{A dataset and data dictionary pair. (a ready4 S4)}

\item{\code{include_chr}}{Include (a character vector)}

\item{\code{label_1L_chr}}{Label (a character vector of length one)}

\item{\code{match_1L_chr}}{Match (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

