% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_CostlySource.R
\docType{class}
\name{CostlySource-class}
\alias{CostlySource-class}
\alias{CostlySource}
\title{CostlySource}
\description{
Input dataset (and metadata) for generating standardised costing datasets.
}
\section{Slots}{

\describe{
\item{\code{Ready4useDyad_r4}}{A dataset and data dictionary pair. (a ready4 S4)}

\item{\code{include_chr}}{Include (a character vector)}

\item{\code{label_1L_chrs}}{(an instance of the character class)}

\item{\code{match_1L_chr}}{Match (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

