% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_corresponding_var}
\alias{get_corresponding_var}
\title{Get corresponding variable}
\usage{
get_corresponding_var(Ready4useDyad_r4, matches_chr, what_1L_chr = "concepts")
}
\arguments{
\item{Ready4useDyad_r4}{A dataset and data dictionary pair. (a ready4 S4)}

\item{matches_chr}{Matches (a character vector)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'concepts'}
}
\value{
Corresponding (a character vector)
}
\description{
get_corresponding_var() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get corresponding variable. Function argument Ready4useDyad_r4 specifies the where to look for the required object. The function returns Corresponding (a character vector).
}
\keyword{internal}
