% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_country_standards}
\alias{get_country_standards}
\title{Get country standards}
\usage{
get_country_standards(
  names_1L_lgl = F,
  default_pkg_ds_chr = c("ISO_3166_1", "ISOcodes"),
  indices_int = c(4:6, 2, 1),
  tbl_index_1L_int = integer(0),
  type_1L_chr = "Countries",
  url_1L_chr = character(0),
  what_1L_chr = "Input_Seed",
  X_Ready4useRepos = ready4use::Ready4useRepos()
)
}
\arguments{
\item{names_1L_lgl}{Names (a logical vector of length one), Default: F}

\item{default_pkg_ds_chr}{Default package dataset (a character vector), Default: c("ISO_3166_1", "ISOcodes")}

\item{indices_int}{Indices (an integer vector), Default: c(4:6, 2, 1)}

\item{tbl_index_1L_int}{Table index (an integer vector of length one), Default: integer(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'Countries'}

\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Input_Seed'}

\item{X_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}
}
\value{
Country standards (an output object of multiple potential types)
}
\description{
get_country_standards() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get country standards. Function argument names_1L_lgl specifies the where to look for the required object. The function returns Country standards (an output object of multiple potential types).
}
\keyword{internal}
