% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_currency}
\alias{get_currency}
\title{Get currency}
\usage{
get_currency(
  country_1L_chr,
  case_when_false_1L_chr = NA_character_,
  case_when_true_1L_chr = NA_character_,
  case_when_true_ls = NULL,
  case_when_var_1L_chr = NA_character_,
  country_var_nms_chr = c("State or territory[1]", "Countries/ territories"),
  currency_tb = NULL,
  filter_cdn_1L_chr = NA_character_,
  force_standard_1L_lgl = F,
  format_1L_chr = "Name",
  fuzzy_logic_1L_chr = "jw",
  indcs_int = c(2:3),
  seed_df = NULL,
  match_var_nm_1L_chr = "ISO code[2]",
  max_distance_1L_dbl = Inf,
  target_var_nm_1L_chr = "Alpha_3",
  tf_false_val_1L_lgl = T,
  type_1L_chr = "Country",
  url_1L_chr = "https://en.wikipedia.org/wiki/List_of_circulating_currencies",
  what_1L_chr = "Symbol",
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  correspondences_y_r3 = ready4show::ready4show_correspondences(),
  X_Ready4useRepos = ready4use::Ready4useRepos()
)
}
\arguments{
\item{country_1L_chr}{Country (a character vector of length one)}

\item{case_when_false_1L_chr}{Case when false (a character vector of length one), Default: 'NA'}

\item{case_when_true_1L_chr}{Case when true (a character vector of length one), Default: 'NA'}

\item{case_when_true_ls}{Case when true (a list), Default: NULL}

\item{case_when_var_1L_chr}{Case when variable (a character vector of length one), Default: 'NA'}

\item{country_var_nms_chr}{Country variable names (a character vector), Default: c("State or territory\link{1}", "Countries/ territories")}

\item{currency_tb}{Currency (a tibble), Default: NULL}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{force_standard_1L_lgl}{Force standard (a logical vector of length one), Default: F}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'Name'}

\item{fuzzy_logic_1L_chr}{Fuzzy logic (a character vector of length one), Default: 'jw'}

\item{indcs_int}{Indices (an integer vector), Default: c(2:3)}

\item{seed_df}{Seed (a data.frame), Default: NULL}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: 'ISO code\link{2}'}

\item{max_distance_1L_dbl}{Maximum distance (a double vector of length one), Default: Inf}

\item{target_var_nm_1L_chr}{Target variable name (a character vector of length one), Default: 'Alpha_3'}

\item{tf_false_val_1L_lgl}{Transform false value (a logical vector of length one), Default: T}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'Country'}

\item{url_1L_chr}{Url (a character vector of length one), Default: 'https://en.wikipedia.org/wiki/List_of_circulating_currencies'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Symbol'}

\item{correspondences_x_r3}{Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()}

\item{correspondences_y_r3}{Correspondences y (a ready4 S3), Default: ready4show::ready4show_correspondences()}

\item{X_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}
}
\value{
Currency (a character vector of length one)
}
\description{
get_currency() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get currency. Function argument country_1L_chr specifies the where to look for the required object. The function returns Currency (a character vector of length one).
}
\keyword{internal}
