% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_currency_tbls}
\alias{get_currency_tbls}
\title{Get currency tables}
\usage{
get_currency_tbls(
  type_1L_chr = c("Both", "Country", "Currency"),
  country_var_nms_chr = c("State or territory[1]", "Countries/ territories"),
  indcs_int = c(2:3),
  url_1L_chr = "https://en.wikipedia.org/wiki/List_of_circulating_currencies",
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  X_Ready4useRepos = ready4use::Ready4useRepos()
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one), Default: c("Both", "Country", "Currency")}

\item{country_var_nms_chr}{Country variable names (a character vector), Default: c("State or territory\link{1}", "Countries/ territories")}

\item{indcs_int}{Indices (an integer vector), Default: c(2:3)}

\item{url_1L_chr}{Url (a character vector of length one), Default: 'https://en.wikipedia.org/wiki/List_of_circulating_currencies'}

\item{correspondences_x_r3}{Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()}

\item{X_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}
}
\value{
Currency table (an output object of multiple potential types)
}
\description{
get_currency_tbls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get currency tables. Function argument type_1L_chr specifies the where to look for the required object. The function returns Currency table (an output object of multiple potential types).
}
\keyword{internal}
