% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_seed_cities}
\alias{get_seed_cities}
\title{Get seed cities}
\usage{
get_seed_cities(
  names_1L_lgl = F,
  indices_int = 2L,
  default_pkg_ds_chr = c("world.cities", "maps"),
  tbl_index_1L_int = integer(0),
  type_1L_chr = "Cities",
  url_1L_chr = character(0),
  what_1L_chr = "Input_Seed",
  X_Ready4useRepos = ready4use::Ready4useRepos()
)
}
\arguments{
\item{names_1L_lgl}{Names (a logical vector of length one), Default: F}

\item{indices_int}{Indices (an integer vector), Default: 2}

\item{default_pkg_ds_chr}{Default package dataset (a character vector), Default: c("world.cities", "maps")}

\item{tbl_index_1L_int}{Table index (an integer vector of length one), Default: integer(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'Cities'}

\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Input_Seed'}

\item{X_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}
}
\value{
Seed cities (an output object of multiple potential types)
}
\description{
get_seed_cities() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get seed cities. Function argument names_1L_lgl specifies the where to look for the required object. The function returns Seed cities (an output object of multiple potential types).
}
\keyword{internal}
