% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_seed_ds}
\alias{get_seed_ds}
\title{Get seed dataset}
\usage{
get_seed_ds(
  label_1L_chr,
  default_pkg_ds_chr = character(0),
  indices_int = integer(0),
  tbl_index_1L_int = integer(0),
  url_1L_chr = character(0),
  what_1L_chr = "Input_Seed",
  X_Ready4useRepos = ready4use::Ready4useRepos()
)
}
\arguments{
\item{label_1L_chr}{Label (a character vector of length one)}

\item{default_pkg_ds_chr}{Default package dataset (a character vector), Default: character(0)}

\item{indices_int}{Indices (an integer vector), Default: integer(0)}

\item{tbl_index_1L_int}{Table index (an integer vector of length one), Default: integer(0)}

\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Input_Seed'}

\item{X_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}
}
\value{
Seed dataset (an output object of multiple potential types)
}
\description{
get_seed_ds() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get seed dataset. Function argument label_1L_chr specifies the where to look for the required object. The function returns Seed dataset (an output object of multiple potential types).
}
\keyword{internal}
