% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_country_correspondences}
\alias{make_country_correspondences}
\title{Make country correspondences}
\usage{
make_country_correspondences(
  option_1L_chr = "custom",
  old_nms_chr = character(0),
  new_nms_chr = character(0)
)
}
\arguments{
\item{option_1L_chr}{Option (a character vector of length one), Default: 'custom'}

\item{old_nms_chr}{Old names (a character vector), Default: character(0)}

\item{new_nms_chr}{New names (a character vector), Default: character(0)}
}
\value{
Correspondences x (a ready4 S3)
}
\description{
make_country_correspondences() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make country correspondences. The function returns Correspondences x (a ready4 S3).
}
\keyword{internal}
