% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_currency_ls}
\alias{make_currency_ls}
\title{Make currency list}
\usage{
make_currency_ls(
  country_1L_chr,
  currency_tb = NULL,
  format_1L_chr = "Name",
  match_var_nm_1L_chr = "ISO code[2]",
  what_1L_chr = "Symbol[D] orAbbrev.[3]"
)
}
\arguments{
\item{country_1L_chr}{Country (a character vector of length one)}

\item{currency_tb}{Currency (a tibble), Default: NULL}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'Name'}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: 'ISO code\link{2}'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Symbol\link{D} orAbbrev.\link{3}'}
}
\description{
make_currency_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make currency list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
