% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_currency_standards}
\alias{make_currency_standards}
\title{Make currency standards}
\usage{
make_currency_standards(as_list_1L_lgl = F, append_ls = NULL)
}
\arguments{
\item{as_list_1L_lgl}{As list (a logical vector of length one), Default: F}

\item{append_ls}{Append (a list), Default: NULL}
}
\value{
Standards (an output object of multiple potential types)
}
\description{
make_currency_standards() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make currency standards. The function returns Standards (an output object of multiple potential types).
}
\keyword{internal}
