% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_defaults}
\alias{make_defaults}
\title{Make defaults}
\usage{
make_defaults(
  label_1L_chr = "Standardised",
  what_1L_chr = c("all", "arguments", "correspondences", "logic", "seed", "reference"),
  force_standard_1L_lgl = F
)
}
\arguments{
\item{label_1L_chr}{Label (a character vector of length one), Default: 'Standardised'}

\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "arguments", "correspondences", "logic", "seed", "reference")}

\item{force_standard_1L_lgl}{Force standard (a logical vector of length one), Default: F}
}
\value{
Default (an output object of multiple potential types)
}
\description{
make_defaults() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make defaults. The function returns Default (an output object of multiple potential types).
}
\keyword{internal}
