% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ds_names}
\alias{make_ds_names}
\title{Make dataset names}
\usage{
make_ds_names(
  label_1L_chr = "Standardised",
  file_nm_1L_lgl = T,
  type_chr = c("Input", "Output"),
  what_chr = c("Correspondences", "Lookup", "Seed", "Standards", "Validation")
)
}
\arguments{
\item{label_1L_chr}{Label (a character vector of length one), Default: 'Standardised'}

\item{file_nm_1L_lgl}{File name (a logical vector of length one), Default: T}

\item{type_chr}{Type (a character vector), Default: c("Input", "Output")}

\item{what_chr}{What (a character vector), Default: c("Correspondences", "Lookup", "Seed", "Standards", "Validation")}
}
\value{
Names (a character vector)
}
\description{
make_ds_names() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dataset names. The function returns Names (a character vector).
}
\keyword{internal}
