% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_standardised_dss}
\alias{make_standardised_dss}
\title{Make standardised datasets}
\usage{
make_standardised_dss(
  label_1L_chr = "Country",
  lookup_Ready4useDyad = ready4use::Ready4useDyad(),
  seed_Ready4useDyad = ready4use::Ready4useDyad(),
  standards_Ready4useDyad = ready4use::Ready4useDyad(),
  type_chr = c("Input", "Output"),
  validation_ls = list(),
  what_chr = c("Correspondences", "Lookup", "Seed", "Standards", "Validation"),
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  correspondences_y_r3 = ready4show::ready4show_correspondences()
)
}
\arguments{
\item{label_1L_chr}{Label (a character vector of length one), Default: 'Country'}

\item{lookup_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{seed_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{standards_Ready4useDyad}{PARAM_DESCRIPTION, Default: ready4use::Ready4useDyad()}

\item{type_chr}{Type (a character vector), Default: c("Input", "Output")}

\item{validation_ls}{Validation (a list), Default: list()}

\item{what_chr}{What (a character vector), Default: c("Correspondences", "Lookup", "Seed", "Standards", "Validation")}

\item{correspondences_x_r3}{Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()}

\item{correspondences_y_r3}{Correspondences y (a ready4 S3), Default: ready4show::ready4show_correspondences()}
}
\value{
Standardised datasets (a list)
}
\description{
make_standardised_dss() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make standardised datasets. The function returns Standardised datasets (a list).
}
\keyword{internal}
