% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_standards_xx}
\alias{make_standards_xx}
\title{Make standards output object of multiple potential types}
\usage{
make_standards_xx(
  as_list_1L_lgl = T,
  append_ls = NULL,
  var_nms_chr = NA_character_,
  label_1L_chr = "Standardised",
  Ready4useDyad_r4 = ready4use::Ready4useDyad()
)
}
\arguments{
\item{as_list_1L_lgl}{As list (a logical vector of length one), Default: T}

\item{append_ls}{Append (a list), Default: NULL}

\item{var_nms_chr}{Variable names (a character vector), Default: 'NA'}

\item{label_1L_chr}{Label (a character vector of length one), Default: 'Standardised'}

\item{Ready4useDyad_r4}{A dataset and data dictionary pair. (a ready4 S4), Default: ready4use::Ready4useDyad()}
}
\value{
Standards (an output object of multiple potential types)
}
\description{
make_standards_xx() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make standards output object of multiple potential types. The function returns Standards (an output object of multiple potential types).
}
\keyword{internal}
