% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_survey}
\alias{make_survey}
\title{Make survey}
\usage{
make_survey()
}
\description{
make_survey() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make survey. The function is called for its side effects and does not return a value.
}
\keyword{internal}
