% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_validation_ls}
\alias{make_validation_ls}
\title{Make validation list}
\usage{
make_validation_ls(allowed_xx, ds_df, var_nm_1L_chr, sort_1L_lgl = F)
}
\arguments{
\item{allowed_xx}{Allowed (an output object of multiple potential types)}

\item{ds_df}{Dataset (a data.frame)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: F}
}
\value{
Validation (a list)
}
\description{
make_validation_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make validation list. The function returns Validation (a list).
}
\keyword{internal}
