% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_validation_output}
\alias{make_validation_output}
\title{Make validation output}
\usage{
make_validation_output(
  allowed_xx,
  ds_df,
  var_nm_1L_chr,
  sort_1L_lgl = F,
  type_1L_chr = "Valid",
  what_1L_chr = "Cases"
)
}
\arguments{
\item{allowed_xx}{Allowed (an output object of multiple potential types)}

\item{ds_df}{Dataset (a data.frame)}

\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: F}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'Valid'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Cases'}
}
\value{
Output (an output object of multiple potential types)
}
\description{
make_validation_output() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make validation output. The function returns Output (an output object of multiple potential types).
}
\keyword{internal}
