% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-CostlyCorrespondences}
\alias{manufacture-CostlyCorrespondences}
\alias{manufacture,CostlyCorrespondences-method}
\alias{manufacture-CostlyCurrencies}
\alias{manufacture,CostlyCurrencies-method}
\alias{manufacture-CostlyCountries}
\alias{manufacture,CostlyCountries-method}
\title{Manufacture a new object}
\usage{
\S4method{manufacture}{CostlyCorrespondences}(
  x,
  sort_1L_lgl = T,
  type_chr = make_ds_names(type_chr = character(0)),
  what_1L_chr = c("dataset", "defaults", "shorthand"),
  ...
)

\S4method{manufacture}{CostlyCurrencies}(
  x,
  sort_1L_lgl = T,
  type_chr = make_ds_names(type_chr = character(0)),
  what_1L_chr = c("dataset", "defaults", "shorthand"),
  ...
)

\S4method{manufacture}{CostlyCountries}(
  x,
  sort_1L_lgl = T,
  type_chr = make_ds_names(type_chr = character(0)),
  what_1L_chr = c("dataset", "defaults", "shorthand"),
  ...
)
}
\arguments{
\item{x}{An object of class CostlyCountries}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: T}

\item{type_chr}{Type (a character vector), Default: make_ds_names(type_chr = character(0))}

\item{what_1L_chr}{What (a character vector of length one), Default: c("dataset", "defaults", "shorthand")}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture method applied to CostlyCorrespondences

manufacture method applied to CostlyCurrencies

manufacture method applied to CostlyCountries
}
