% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure-CostlySource}
\alias{procure-CostlySource}
\alias{procure,CostlySource-method}
\title{Procure items from a dataset}
\usage{
\S4method{procure}{CostlySource}(
  x,
  matches_chr = character(0),
  what_1L_chr = c("names", "concepts"),
  ...
)
}
\arguments{
\item{x}{An object of class CostlySource}

\item{matches_chr}{Matches (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("names", "concepts")}

\item{...}{Additional arguments}
}
\value{
Object (an output object of multiple potential types)
}
\description{
procure method applied to CostlySource
}
