% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ratify.R
\name{ratify-CostlyCorrespondences}
\alias{ratify-CostlyCorrespondences}
\alias{ratify,CostlyCorrespondences-method}
\title{Ratify that a dataset meets validity criteria}
\usage{
\S4method{ratify}{CostlyCorrespondences}(
  x,
  new_val_xx = "identity",
  sort_1L_lgl = T,
  type_1L_chr = make_ds_names(type_chr = character(0))[1:2],
  what_chr = make_ds_names(type_chr = character(0)),
  ...
)
}
\arguments{
\item{x}{An object of class CostlyCorrespondences}

\item{new_val_xx}{New value (an output object of multiple potential types), Default: 'identity'}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: T}

\item{type_1L_chr}{Type (a character vector of length one), Default: make_ds_names(type_chr = character(0))\link{1:2}}

\item{what_chr}{What (a character vector), Default: make_ds_names(type_chr = character(0))}

\item{...}{Additional arguments}
}
\value{
x (An object of class CostlyCorrespondences)
}
\description{
ratify method applied to CostlyCorrespondences
}
