% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew-CostlyCorrespondences}
\alias{renew-CostlyCorrespondences}
\alias{renew,CostlyCorrespondences-method}
\title{Renew values in a dataset}
\usage{
\S4method{renew}{CostlyCorrespondences}(
  x,
  new_val_xx = NULL,
  type_1L_chr = "slot",
  what_1L_chr = "correspondences",
  ...
)
}
\arguments{
\item{x}{An object of class CostlyCorrespondences}

\item{new_val_xx}{New value (an output object of multiple potential types), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'slot'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'correspondences'}

\item{...}{Additional arguments}
}
\value{
x (An object of class CostlyCorrespondences)
}
\description{
renew method applied to CostlyCorrespondences
}
