% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_country}
\alias{transform_country}
\title{Transform country}
\usage{
transform_country(
  country_1L_chr,
  from_1L_chr = "Name",
  standards_df = data.frame(),
  what_1L_chr = "Alpha_3"
)
}
\arguments{
\item{country_1L_chr}{Country (a character vector of length one)}

\item{from_1L_chr}{From (a character vector of length one), Default: 'Name'}

\item{standards_df}{Standards (a data.frame), Default: data.frame()}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Alpha_3'}
}
\value{
Country (a character vector of length one)
}
\description{
transform_country() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform country. Function argument country_1L_chr specifies the object to be updated. Argument from_1L_chr provides the object to be updated. The function returns Country (a character vector of length one).
}
\keyword{internal}
