% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_country_default_ls}
\alias{update_country_default_ls}
\title{Update country default list}
\usage{
update_country_default_ls(default_ls, force_standard_1L_lgl = F)
}
\arguments{
\item{default_ls}{Default (a list)}

\item{force_standard_1L_lgl}{Force standard (a logical vector of length one), Default: F}
}
\value{
Default (a list)
}
\description{
update_country_default_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update country default list. Function argument default_ls specifies the object to be updated. Argument force_standard_1L_lgl provides the object to be updated. The function returns Default (a list).
}
\keyword{internal}
