% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_currency_correspondences}
\alias{update_currency_correspondences}
\title{Update currency correspondences}
\usage{
update_currency_correspondences(
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  country_var_nms_chr = c("State or territory[1]", "Countries/ territories"),
  type_1L_chr = c("Both", "Country", "Currency")
)
}
\arguments{
\item{correspondences_x_r3}{Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()}

\item{country_var_nms_chr}{Country variable names (a character vector), Default: c("State or territory\link{1}", "Countries/ territories")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("Both", "Country", "Currency")}
}
\value{
Correspondences x (a ready4 S3)
}
\description{
update_currency_correspondences() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update currency correspondences. Function argument correspondences_x_r3 specifies the object to be updated. Argument country_var_nms_chr provides the object to be updated. The function returns Correspondences x (a ready4 S3).
}
\keyword{internal}
