% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_module_slot}
\alias{update_module_slot}
\title{Update module slot}
\usage{
update_module_slot(x_r4, y_r3, what_1L_chr, new_val_xx = NULL)
}
\arguments{
\item{x_r4}{X (a ready4 S4)}

\item{y_r3}{Y (a ready4 S3)}

\item{what_1L_chr}{What (a character vector of length one)}

\item{new_val_xx}{New value (an output object of multiple potential types), Default: NULL}
}
\value{
X (a ready4 S4)
}
\description{
update_module_slot() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update module slot. Function argument x_r4 specifies the object to be updated. Argument y_r3 provides the object to be updated. The function returns X (a ready4 S4).
}
\keyword{internal}
