% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_with_standards}
\alias{update_with_standards}
\title{Update with standards}
\usage{
update_with_standards(
  seed_df,
  standards_df,
  standards_var_nms_chr,
  correspondences_x_r3 = ready4show::ready4show_correspondences(),
  case_when_false_1L_chr = NA_character_,
  case_when_true_1L_chr = NA_character_,
  case_when_true_ls = NULL,
  case_when_var_1L_chr = NA_character_,
  order_1L_lgl = F,
  reference_var_nm_1L_chr = "country.etc",
  filter_cdn_1L_chr = NA_character_,
  force_standard_1L_lgl = F,
  fuzzy_logic_1L_chr = character(0),
  max_distance_1L_dbl = Inf,
  tf_false_val_1L_lgl = T
)
}
\arguments{
\item{seed_df}{Seed (a data.frame)}

\item{standards_df}{Standards (a data.frame)}

\item{standards_var_nms_chr}{Standards variable names (a character vector)}

\item{correspondences_x_r3}{Correspondences x (a ready4 S3), Default: ready4show::ready4show_correspondences()}

\item{case_when_false_1L_chr}{Case when false (a character vector of length one), Default: 'NA'}

\item{case_when_true_1L_chr}{Case when true (a character vector of length one), Default: 'NA'}

\item{case_when_true_ls}{Case when true (a list), Default: NULL}

\item{case_when_var_1L_chr}{Case when variable (a character vector of length one), Default: 'NA'}

\item{order_1L_lgl}{Order (a logical vector of length one), Default: F}

\item{reference_var_nm_1L_chr}{Reference variable name (a character vector of length one), Default: 'country.etc'}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{force_standard_1L_lgl}{Force standard (a logical vector of length one), Default: F}

\item{fuzzy_logic_1L_chr}{Fuzzy logic (a character vector of length one), Default: character(0)}

\item{max_distance_1L_dbl}{Maximum distance (a double vector of length one), Default: Inf}

\item{tf_false_val_1L_lgl}{Transform false value (a logical vector of length one), Default: T}
}
\value{
Standardised dataset (a data.frame)
}
\description{
update_with_standards() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update with standards. Function argument seed_df specifies the object to be updated. Argument standards_df provides the object to be updated. The function returns Standardised dataset (a data.frame).
}
\keyword{internal}
